Add files via upload
ABSTRACT

Everyone has a dream to buy a new car. But they cannot afford to buy a new car at that time he wants to buy a used cars. They do not have a knowledge to predict the price of a car. At that time this project will be helpful to predict the car selling price based on their features like the car model, the number of years that a car is old, the type of fuel it uses, the type of seller, the type of transmission and the number of kilometers that the car has driven so far. This project will help to get an approximation about selling price of a used car based on its features and reduces the seller and consumer risk in business. The proposed model utilizes the machine learning algorithms and regression techniques of statistics decision tree, XG boost and random forest regressions to achieve this task.

 ALGORITHM:

XGBoost is a robust machine-learning algorithm that can help you understand your data and make better decisions. XGBoost is an implementation of gradient-boosting decision trees. It has been used by data scientists and researchers worldwide to optimize their machine-learning models. XGBoost is designed for speed, ease of use, and performance on large datasets. 
Initialization:
1.	Set an initial prediction (usually the mean of the target variable).
2.	Calculate the initial residuals.
For Each Iteration:
1.	Train a weak learner (decision tree) on the negative gradients of the loss function.
2.	Calculate the learning rate-adjusted contribution of the tree to the model.
3.	Update predictions and residuals.
Regularization:
1.	Apply L1 and L2 regularization to control the complexity of the model.
Stopping Criteria:
1.	Determine when to stop adding trees based on a user-defined number of iterations or when no significant improvement is observed.